UE5.2.1 - Collapsible UI
Revised the MMVR Editor Utility Widget to have collapsible subcategories, which will allow more useful features and better organization.

Inside the MMVR Editor Utility Widget is the 'Tracker Adjustment' field, which shows your tracker layout at a glance, and allows loading and saving a preset to your player pawn.  Clicking the buttons when in actor mode will freeze that goal for an alotted time, allowing you to move individual parts into better alignment.

Revised the hand IK rig used with leap motion to get more shoulder motion and to fix a twisting issue I noticed on the left upper arm.

Fixed up some errors in the light rig where the 'arrow' component (the base of all the lights) wasn't moving to the midpoint of the actor when not in play mode.  Now the light rig should act more similar across play mode and when using the editor outside of play mode.

Made the light rig editor a vertical arrangement, which should make it easier to work on cutscenes while having the editor utility open.