UE5.2.1 - Revised Foot IK
Changed the player pawn to counterrotate the IK retargeting actor in calibration so that you can change the player pawn's relative rotation to the character.

Increased sample count on the rect light in the light rig.

Completely revised the foot IK system, and it should actually lock to world space now, as well as have proper tilt in most environments.

Fixed light transition time in the light rig.

made the Spectator camera focal length more reflective of its defaults in game.

Added calculate pole vector to several control rigs, as well as a soft IK for just the legs.  It might keep the feet a few inches in the air, but will avoid the popping that happens when the IK chain reaches its maximum length.

Nearly finished a new finger solver that is in the WIP bin.  Unfortunately it seems to break during use in cinematics/animation, but if anyone could get this working it would be great!  It possibly could be better suited to runtime/vr applications.