UE5.3.2 - Upper Body Solve Improvements
Removed the legacy input system, and fixed up some bugs with the thumbstick not triggering events properly.  Removed the input ini file.

Improved the upper body solve on the UE5 many and Metahuman set up with a new function that enables clav movement and then locked local rotations to a certain range.  Lower Arm popping should be reduced at most angles.

Swapped the left and right sides on the Tracker Offset Widget in the Editor Utility Widget.  Added a Save Button to save any offsets made to the tracker offsets.  Added a calibration timer to the to the calibration section: entering a number and hitting the button will calibrate after a delay.

Deprecated the facial expressions additives.