UE5.3.2 - MMVR Main Menu
Started making a general User UI for standalone builds.  In game, pressing the tab key will toggle a menu that has a few controls set up (not all controls are currently hooked up however).  If you level isn't using the MMVR game mode, tab might not toggle the mouse cursor to be visible.

Moved the stretch sense anim bp to the 'add-ons' folder to make it easier for modification if the hand engine plugin isn't installed.

Ensured that the MMVR, Metahuman, CC3 and UE4 control rigs bake the fingers in the backwards solve.  Metahuman heads no longer align with the bodies after solving, which is annoying, but I haven't been able to figure that out.  Revised CC3 control rig to have limits on the clav movement.

Added some properties to the light rig that makes it load and save more consistently.

Changed the example level to use the MMVR game base: Added some functionality to show the mouse cursor in that game mode.