UE5.3.2 - Manny Movement Actor Update
Changed the order and operation during begin play for spawning in player controllers.  Made the player pawn not default to auto-posess the player, as this is now handled in begin play.  Disabled split screen when multiple player pawns are present in scene.  In general, tested local multiplayer and it should work better now.

Completely redid the movement actor to use a UE5 mannequin.  Not everything works like the old set up, but the animations should be better quality, and eventually it will be easier to extend in terms of additional animations and motion matching features in the future.  Using the movement actor in VR pie mode will offset the player pawn to match the capsule for first person gameplay.

Removed some collision from the new spectator camera mesh.

Rotated the VR Rig -90 in the player pawn so it would match the rotation that actors are set up with in their blueprints.  This seems to have helped a lot of system, but might have some knock on effects.  please let me know if you find any!

Added gamepad inputs to the enhanced input mapping  context, so you can move your actor around with the gamepad.

Made it so that the spectator camera doesn't auto grab input from player 0, but will use the player controller that you set as the public int.  This should allow for more flexibility in the future.   Made it so the spectator fly cam can be used at any point even if not being driven by a tracker.