UE5.3.2 - OpenXR VR Menu
Added motion controllers back into the player pawn for the OpenXR tracking mode: they are set to be disabled by default unless the player choses that mode AND launches the game in VR.  I found there are occasionally crashes and instabilities on exiting playmode, but overall this method is a lot more stable and versatile than previous implementations of OpenXR, which could crash when run with the -XRtrackingonly flag on startup.

Changed the player pawn settings to be more consistent when using the system with VR: now it should always render the spectator camera view when in VR if that is present in the scene as long as you check the 'in HMD capture' bool on the player pawn.

Added a bool to disable gamepad input on the Spectator camera and the player pawn.

Noticed some bugs were introduced when capturing in VR using a non movement actor, so I introduced a function that will rotate the VR Rig back to its original settings when working in VR.

Made an in-game VR menu with widget interaction for adjusting settings when in a build/hmd.

Revised the movement actor to get more consistent behavior, however upper body mode is currently breaking the set up.  Also, in VR, this set up is causing some shaking, so it might be better to avoid using the hmd capture mode with the movement actor for now.

Renamed and reorganized, a lot of components, which caused nearly every blueprint to be recompiled.  I fixed redirectors, and hopefully nothing is missing.