UE5.3.2 - Spectator Camera Rework
Added a new CineCamera BP derived from the engine cinecamera class into the project.  When creating a new camera from the MMVR_SpectatorCamera, this new class will spawn, allowing you to set custom film back/postprocessing stacks for your virtual cameras.

Added a new BP_CameraTarget which spawns when creating a new camera.  It will attempt to attach itself to the current mocap actor, and should record into takes/make for an easy focus point or position to track.  NOTE: When making cut scenes, its best if these actors exist in your level and are NOT spawned in when you create a level sequence as the binding often fails.

Revised the autofocus system on the spectator camera, averaging a series of traces.  Added a new camera shake preset.

Added a 'detatch camera' button to the editor utility widget, and fixed up functionality in the 'reset cam tracker' so now it freezes the current camera position until pressed again.  This functionality can also be accessed with the 'start' button on the controller.

Stopped the trigger from scaling when the VR Menu is open (you can open this by hold both 'B' Buttons on the controllers).  Also fixed up a bug in the calibration screen overlay not disappearing/showing a count down timer.