UE5.3.2 - VR Rework
Various changes to make VR a better experience.  To capture in VR, it's best to import MMVR into a VR template project, change the player pawn to OpenXR tracking mode and turn on the bool for HMD capture.

Fixed up the stretching fingers if the left hand got too far away from the torso on the UE5 Manny control rig.  In the UE5 control rig, setting the controls to the fingers messes up the finger rotation later for the hand control rig, so I disconnected those.

Changed the CC3 spine so it shouldn't stretch anymore.  Also fixed UE4 control rigs to work with upper body only mode.

Fixed a bug where metahumans weren't working with audio based mouth movements.

Changed the spectator camera to not spawn in a BP_CameraTarget when making an external camera.  Its best to bring this into your scenes and attach it to actors in sequencer as needed.