UE5.3.2 - Actor List
Changed OSC Receiver to also get signals from Self Reflect, however I was unable to figure out the quaternion rotations so none of the bones comes through correctly.

Added a VRM skeleton for debugging purposes.

Changed the Function Add Player Pawn Rotation to offset the VR player if using IK retargeting.

Created an actor list in the editor utility widget that will search through your scene for mocap actors and give you a handy panel to adjust properties on them, such as if they are the mocap target or not.

Extended Player Pawn settings panel in the VRMenu.

Made light rig editor default to UE5 bone names to allow updating when not in play mode.  Added the same bug fix to allow the camera to be framed on actors when not in play mode.  Fixed some bugs in Light Rig not setting proper intensity.