UE5.4.4 - MMVR Cine Camera Update
Revised the MMVR Cine Camera Actor to set a tracking look at actor and focus tracking based on a subject.

Added that simplified camera to the new scene builder, although it doesn't work well in 5.4 when recording to sequencer as spawnable actors.  If you want to use this system, possessable might work better!