UE5.4.4 - Puppet Actor
Added a new 'puppet' character set up for driving simple characters.  Most of the code was written with the intension of being rig agnostic, so it can also work with other meshes as long as you duplicate the Puppet Animation Blueprint and assign the skeleton of whatever character you are working with.

To use the puppet blueprint, put one tracker on the inside of your forearm bear the elbow, and one tracker on the top of your wrist.  In Pie play mode, lower your hand as if puppeteering a muppet, and hit the calibrate button on the bp (or set up the 'Calibrate after X seconds' variable).

To make your own puppet, make a child bp of the main puppet bp and duplicate it's animation bp, reassigning it's mesh to the one you want to use.  Then update your child bp with the proper mesh, the newly created animation bp, as well as the bone names/jaw/eyelid rotations of your mesh).

Cleaned up some file names on the float curves.

Fixed up some spawning on the scene builder editor utility widget.