UE5.6.1 - OSC Bug Fix
Fixed a bug with the OSC receiver: after UE5.5, making an OSC server with 127.0.0.1 to receive local messages from the same machine didn't work, so I switched that to 0.0.0.0

Removed the iPhoneOnly bool on the player pawn, replacing it with an option in the body tracking Enum for IphoneOnly.

Modified the OSC receiver to get VRChat eye tracking information from the big screen beyond 2e.  All data should be sent over additively, although it can override eye bone rotations if an iphone is also being used on the character.

Organized the public variables on the player pawn a bit, seperating out calibration specific variables.

Added eye movement and jaw movement to UE4 set up.

Metahuman's don't currently work with metahumans, but I am looking into it.