UE5.6.1 - Metahuman Animator Update
Updated the system to work with the new Metahuman Animator system in 5.6.

Now, all you need to do to your metahumans is add the mocap actor component, set it to metahuman, rotate the body -90 degrees and add the MMVR_metahuman_animbp to the body mesh.

The biggest change is I moved most of the nodes from the metahuman animbp to the body animbp since that is using set leader pose to control the head now.

The system should detect if you are using metahuman animator as the live link subject or an iphone.  Please note you might have to disable sending head motion if you are using a head mounted camera.

Deleted the iphone only mode bool from some menus as this functionality is now in the body tracking enum, and removed the deprecated metahuman body control rig.